devices: Remove superfluous exception handling
We shouldn't be handling MbedTargetsErrors in our get_connected_device
function as it breaks encapsulation. This function only knows that a
device will be created, added to the ConnectedDevices and the
ConnectedDevices will be returned. This function does not care about
platform database errors. We should just allow the exceptions to bubble
up as they will contain more relevant information as to the cause of the
errors in their error messages compared to the information we can
provide here. Also we now no longer have a layering violation.